% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveil.R
\name{as_thrive}
\alias{as_thrive}
\alias{as_thrive.survfit}
\alias{as_thrive.numeric}
\title{Coerce a fit to a thrive object}
\usage{
as_thrive(x, ...)

\method{as_thrive}{survfit}(x, ...)

\method{as_thrive}{numeric}(x, ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Coerce a fit to a thrive object
}
